requirements.txt
openai
aiogram
python-dotenv
tinydb==4.8.2
cryptography==43.0.1